pyproject.toml
[project]
name = "identity-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.1",
    "celery>=5.6.2",
    "cryptography>=46.0.5",
    "fastapi>=0.128.0",
    "flower>=2.0.1",
    "httpx>=0.28.1",
    "jwcrypto>=1.5.6",
    "passlib[argon2]>=1.7.4",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
